requirements.txt
Django
djangorestframework
psycopg2
drf_spectacular
Pillow
uwsgi